pyproject.toml
[tool.poetry]
name = "linkedin_jobs_search_helper"
version = "0.1.0"
description = ""
authors = ["Strong Punch Xavier"]
packages = [{ include = "linkedin_jobs_search_helper", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "4.9.0"
webdriver-manager = ">=4.0.2"
plum-dispatch = "^2.1.1"
pyyaml = "^6.0"
tiktoken = "^0.4.0"
regex = "^2023.6.3"
pydantic = "2.4.0"
openai = "^0.27.8"
colorlog = "^6.7.0"
langid = "^1.1.6"
pymongo = "^4.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"

[tool.pytest.ini_options]
pythonpath = ["src"]
norecursedirs = ["docker/mongo/data"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
